Authentification

+++ Controllers/HotelController.cs
@@ -1,3 +1,5 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using viesbuciu_rezervacija_backend.Data;
@@ -38,33 +40,43 @@ namespace viesbuciu_rezervacija_backend.Controllers
             return Ok(hotel.ToHotelDto());
         }
         [HttpPost]
+        [Authorize(Roles = "HotelOwner")]
         public async Task<IActionResult> PostHotel([FromBody] CreateHotelRequestDto hotelDto)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             var hotelModel = hotelDto.ToHotelFromCreateDto();
+            hotelModel.OwnerId = userId;
             await _hotelRepo.CreateAsync(hotelModel);
             return CreatedAtAction(nameof(GetById), new { id = hotelModel.Id }, hotelModel);
         }
 
         [HttpPut("{id}")]
+        [Authorize(Roles = "HotelOwner")]
         public async Task<IActionResult> UpdateHotel([FromRoute] int id, [FromBody] UpdateHotelRequestDto hotelDto)
         {
-            var hotelModel = await _hotelRepo.UpdateHotel(id, hotelDto);
-            if (hotelModel == null)
-            {
-                return NotFound();
-            }
-            return Ok(hotelModel.ToHotelDto());
+            var hotel = await _hotelRepo.GetByIdAsync(id);
+            if (hotel == null)
+                return NotFound("Hotel not found");
+
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (hotel.OwnerId != userId)
+                return StatusCode(StatusCodes.Status403Forbidden, "You can only update your own Hotel.");
+            var updatedHotel = await _hotelRepo.UpdateHotel(id, hotelDto);
+            return Ok(updatedHotel.ToHotelDto());
         }
 
         [HttpDelete("{id}")]
+        [Authorize(Roles = "HotelOwner")]
         public async Task<IActionResult> DeleteHotel([FromRoute] int id)
         {
-            var hotelModel = await _hotelRepo.DeleteHotelAsync(id);
-            if (hotelModel == null)
-            {
-                return NotFound();
-            }
+            var hotel = await _hotelRepo.GetByIdAsync(id);
+            if (hotel == null)
+                return NotFound("Hotel not found");
 
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (hotel.OwnerId != userId)
+                return StatusCode(StatusCodes.Status403Forbidden, "You can only delete your own Hotel.");
+            await _hotelRepo.DeleteHotelAsync(id);
             return NoContent();
         }
     }
